+++ lib/Output/EPubReader.php
@@ -83,7 +83,7 @@ class EPubReader
     public static function getContent($idData, $component, $request)
     {
         /** @var Book */
-        $book = Book::getBookByDataId($idData);
+        $book = Book::getBookByDataId($idData, $request->database());
         $params = ['data' => $idData, 'db' => $book->getDatabaseId()];
 
         $epub = new static::$epubClass($book->getFilePath('EPUB', $idData));
@@ -104,7 +104,7 @@ class EPubReader
      */
     public static function getReader($idData, $request)
     {
-        $book = Book::getBookByDataId($idData);
+        $book = Book::getBookByDataId($idData, $request->database());
         $params = ['data' => $idData, 'db' => $book->getDatabaseId()];
 
         try {